Added AdbDevice APIs needed to execute the boot event.
Added 6 new APIs to device.py, implementing functionality
for pushing a file to a device, enabling trace on boot on
a device, performing a boot on a device, running the
command, ‘adb wait-for-device’, on a device, checking if a
boot is completed on a device, and waiting for a boot to
complete on a device. Furthermore, added 14 new
units tests to and modified a helper method in
test/device_unit_test.py to test the new APIs.

Test: atest device_unit_test
Bug: 357914823

Change-Id: I97da5e243d2695bc038256a9e8336335cd80c8a0